requirements.txt
click==8.0.3
Flask==2.0.2
fpdf==1.7.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pdf2image==1.16.0
Pillow==8.4.0
Werkzeug==2.0.2
